+++ core/main.ts
@@ -14,8 +14,10 @@ export * from "./src/contrib/systems/core/PhysicsSystem";
 export * from "./src/contrib/systems/gameplay/EnemySpawnSystem";
 export * from "./src/contrib/systems/gameplay/PlayerSpawnSystem";
 export * from "./src/contrib/systems/net/TickData";
-export * from "./src/contrib/systems/net/WsClientSystem";
-export * from "./src/contrib/systems/net/WsServerSystem";
+export * from "./src/contrib/systems/net/delay/DelayClientSystem";
+export * from "./src/contrib/systems/net/delay/DelayServerSystem";
+export * from "./src/contrib/systems/net/rollback/RollbackClientSystem";
+export * from "./src/contrib/systems/net/rollback/RollbackServerSystem";
 export * from "./src/contrib/systems/ui/ClickableSystem";
 export * from "./src/contrib/systems/ui/DebugSystem";
 export * from "./src/contrib/systems/ui/HealthBarSystem";
