+++ core/src/contrib/systems/net/TickData.ts
@@ -4,9 +4,20 @@ export type TickData = {
   type: "game"
   tick: number
   timestamp: number
+  latency?: number
   player: string
   actions: Record<number, Record<string, string[]>>
   chats: Record<number, Record<string, string[]>>
   serializedEntities: Record<string, SerializedEntity>
+}
+
+export type DelayTickData = {
+  type: "game"
+  tick: number
+  timestamp: number
   latency?: number
+  player: string
+  actions: Record<string, string[]>
+  chats: Record<string, string[]>
+  serializedEntities: Record<string, SerializedEntity>
 }
